[Refactor] Move storage primitives into dedicated modules
