+++ be/src/storage/index/inverted/clucene/clucene_inverted_writer.h
@@ -24,8 +24,8 @@
 
 #include "storage/index/inverted/inverted_index_option.h"
 #include "storage/index/inverted/inverted_writer.h"
-#include "storage/key_coder.h"
 #include "storage/olap_common.h"
+#include "storage/primitive/key_coder.h"
 #include "storage/primitive/rowid_types.h"
 #include "storage/tablet_schema.h"
