+++ be/src/storage/index/vector/tenann/del_id_filter.cpp
@@ -36,7 +36,7 @@
 #include "storage/index/vector/tenann/del_id_filter.h"
 
 #include "storage/primitive/range.h"
-#include "storage/roaring2range.h"
+#include "storage/primitive/roaring2range.h"
 #include "tenann/common/type_traits.h"
 #include "tenann/searcher/id_filter.h"
