+++ be/src/storage/primitive/key_coder.cpp
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "storage/key_coder.h"
+#include "storage/primitive/key_coder.h"
 
 namespace starrocks {
