+++ be/src/storage/primitive/key_coder.h
@@ -41,7 +41,6 @@
 #include "common/status.h"
 #include "gutil/strings/substitute.h"
 #include "runtime/mem_pool.h"
-#include "storage/types.h"
 #include "types/datum.h"
 #include "types/storage_type_traits.h"
