+++ be/src/storage/rowset/bitmap_index_evaluator.cpp
@@ -17,7 +17,7 @@
 #include "column/chunk_schema_helper.h"
 #include "common/config_rowset_fwd.h"
 #include "storage/predicate_tree/predicate_tree.hpp"
-#include "storage/roaring2range.h"
+#include "storage/primitive/roaring2range.h"
 
 namespace starrocks {
