+++ be/src/storage/rowset/ordinal_page_index.cpp
@@ -39,7 +39,7 @@
 #include "common/logging.h"
 #include "fs/fs.h"
 #include "runtime/exec_env.h"
-#include "storage/key_coder.h"
+#include "storage/primitive/key_coder.h"
 #include "storage/rowset/page_handle.h"
 #include "storage/rowset/page_io.h"
