+++ be/src/storage/rowset/rowset_writer.h
@@ -44,8 +44,8 @@
 #include "gutil/macros.h"
 #include "runtime/global_dict/types.h"
 #include "runtime/global_dict/types_fwd_decl.h"
-#include "storage/column_mapping.h"
 #include "storage/compaction_utils.h"
+#include "storage/primitive/column_mapping.h"
 #include "storage/rows_mapper.h"
 #include "storage/rowset/rowset.h"
 #include "storage/rowset/rowset_writer.h"
