+++ be/src/storage/rowset/segment_iterator.cpp
@@ -62,9 +62,9 @@
 #include "storage/primitive/chunk_iterator.h"
 #include "storage/primitive/projection_iterator.h"
 #include "storage/primitive/range.h"
+#include "storage/primitive/roaring2range.h"
 #include "storage/primitive/rowid_types.h"
 #include "storage/primitive/vector_search_option.h"
-#include "storage/roaring2range.h"
 #include "storage/rowset/bitmap_index_evaluator.h"
 #include "storage/rowset/bitmap_index_reader.h"
 #include "storage/rowset/column_decoder.h"
