+++ be/src/tools/meta_tool.cpp
@@ -76,11 +76,11 @@
 #include "storage/chunk_helper.h"
 #include "storage/data_dir.h"
 #include "storage/delta_column_group.h"
-#include "storage/key_coder.h"
 #include "storage/lake/tablet_manager.h"
 #include "storage/lake/vacuum.h"
 #include "storage/olap_define.h"
 #include "storage/primary_key_dump.h"
+#include "storage/primitive/key_coder.h"
 #include "storage/primitive/storage_stats.h"
 #include "storage/primitive/zone_map_detail.h"
 #include "storage/rowset/binary_plain_page.h"
