+++ be/test/storage/rowset/bitmap_index_test.cpp
@@ -46,8 +46,8 @@
 #include "runtime/mem_pool.h"
 #include "runtime/mem_tracker.h"
 #include "storage/chunk_helper.h"
-#include "storage/key_coder.h"
 #include "storage/olap_common.h"
+#include "storage/primitive/key_coder.h"
 #include "storage/rowset/bitmap_index_reader.h"
 #include "storage/rowset/bitmap_index_writer.h"
 #include "storage/types.h"
