+++ be/test/storage/rowset/bloom_filter_index_reader_writer_test.cpp
@@ -39,8 +39,8 @@
 #include "common/logging.h"
 #include "fs/fs_memory.h"
 #include "runtime/mem_tracker.h"
-#include "storage/key_coder.h"
 #include "storage/olap_common.h"
+#include "storage/primitive/key_coder.h"
 #include "storage/rowset/bloom_filter_index_reader.h"
 #include "storage/rowset/bloom_filter_index_writer.h"
 #include "storage/types.h"
