+++ be/test/storage_base/CMakeLists.txt
@@ -21,6 +21,7 @@ set(STORAGE_BASE_TEST_FILES
     ${BE_TEST_MACOS_LINK_STUBS}
     merge_iterator_test.cpp
     row_source_mask_test.cpp
+    short_key_index_test.cpp
 )
 
 set(STORAGE_BASE_TEST_LINK_LIBS
