+++ be/test/storage_base/row_source_mask_test.cpp
@@ -19,7 +19,7 @@
 #include "common/config_rowset_fwd.h"
 #include "common/config_storage_fwd.h"
 #include "fs/fs_util.h"
-#include "storage/base/storage_path_constants.h"
+#include "storage/primitive/storage_path_constants.h"
 
 namespace starrocks {
