+++ be/test/storage_primitive/key_coder_test.cpp
@@ -32,7 +32,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-#include "storage/key_coder.h"
+#include "storage/primitive/key_coder.h"
 
 #include <gtest/gtest.h>
