+++ be/test/storage_primitive/roaring2range_test.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "storage/roaring2range.h"
+#include "storage/primitive/roaring2range.h"
 
 #include <gtest/gtest.h>
